requirements.txt
anyio==3.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
google-auth==2.16.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
oauthlib==3.2.2
pyasn1==0.4.8
python-dotenv==0.19.2
pyasn1-modules==0.2.8
python-telegram-bot==20.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
urllib3==1.26.14